Better error message for backoff provided
git-svn-id: file:///dev/shm/somefilter.svn@466 e102df66-1e2e-11dd-9b44-c24451a4db5e

+++ lm/read_arpa.hh
@@ -54,7 +54,7 @@ template <class Voc, class Weights> void ReadNGram(util::FilePiece &f, const uns
     }
     ReadBackoff(f, weights);
   } catch(util::Exception &e) {
-    e << " in the " <<  n << "-gram at byte " << f.Offset();
+    e << " in the " << static_cast<unsigned int>(n) << "-gram at byte " << f.Offset();
     throw;
   }
 }
